feat: 회사 지원금 등록/수정 추가, 사용자 권한 관리 수정

+++ domain/src/main/kotlin/kr/co/jsol/domain/device/entity/Device.kt
@@ -38,7 +38,7 @@ class Device(
     var series: String,
 
     @Comment("가격")
-    var price: Int,
+    var price: Long,
 
     @Comment("용량 ex) 125GB, 256TB 등..")
     var volume: String,
